chore(ZMS): rename mail cron jobs to queue writers
These jobs write to the mail queue; zmsmessaging sends. Also drop the leftover cron__sendNotificationReminder config key.
