+++ zmsbackend/bin/queueMailReminder
@@ -1,0 +1,32 @@
+#!/usr/bin/env php
+<?php
+require_once(__DIR__."/script_bootstrap.php");
+setlocale (LC_ALL, "de_DE.UTF-8");
+
+$usage = <<<EOS
+
+Usage: {$argv[0]} minutes [--verbose] --commit
+        ATTENTION! Queues email reminders based on process reminder timestamp. USE WITH CAUTION!
+        minutes         minutes before appointment to queue email reminder
+        --commit        no dry run, write mails to the queue
+        --verbose       show what would be queued
+
+EOS;
+
+$defaultTimeInterval = preg_grep('#^\d+$#', $argv);
+$verbose = preg_grep('#^--?v(erbose)?$#', $argv);
+$commit = preg_grep('#^--commit$#', $argv);
+
+$now = new \DateTimeImmutable();
+if (class_exists('\App') && isset(\App::$now)) {
+    $now = \App::$now;
+}
+if ($defaultTimeInterval) {
+    $job = new \BO\Zmsbackend\Helper\QueueMailReminder($now, array_shift($defaultTimeInterval), $verbose);
+    $job->startProcessing($commit);
+} else {
+    \App::$log->error('Missing default time interval argument', ['usage' => trim($usage)]);
+if (! $commit) {
+    \App::$log->notice('Use with --commit to queue emails.');
+}
+}
