+++ zmsbackend/bin/queueProcessListToScopeAdmin
@@ -1,0 +1,18 @@
+#!/usr/bin/env php
+<?php
+require_once(__DIR__."/script_bootstrap.php");
+
+$scopeIds = preg_grep('#^\d+$#', $argv);
+$scopeId = false;
+if (count($scopeIds)) {
+    $scopeId = array_shift($scopeIds);
+}
+$verbose = preg_grep('#^--?v(erbose)?$#', $argv);
+$commit = preg_grep('#^--commit$#', $argv);
+
+$job = new \BO\Zmsbackend\Helper\QueueProcessListToScopeAdmin($verbose, $scopeId);
+$job->startProcessing($commit);
+
+if (!$commit) {
+    \App::$log->notice('Use with --commit to queue emails.');
+}
