+++ zmsbackend/composer.json
@@ -100,8 +100,8 @@
         "bin/resetApiQuota",
         "bin/resetGhostWorkstationCount",
         "bin/resetWorkstations",
-        "bin/sendMailReminder",
-        "bin/sendProcessListToScopeAdmin",
+        "bin/queueMailReminder",
+        "bin/queueProcessListToScopeAdmin",
         "bin/updateDldbData"
     ],
     "autoload": {
