+++ zmsbackend/cron/cronjob.daily
@@ -26,7 +26,7 @@ if [[ $ZMS_CRONROOT =~ ^1|on|true|yes$ ]]; then
         $API config cron deleteDayoffData|grep -q "$ZMS_ENV" && $BIN/deleteDayoffData 6 --commit
         $API config cron deleteBlacklistedMail|grep -q "$ZMS_ENV" && $BIN/deleteBlacklistedMail --commit
         $API config cron deleteOldAvailabilityData|grep -q "$ZMS_ENV" && $BIN/deleteOldAvailabilityData 28 --commit
-        $API config cron sendProcessListToScopeAdmin|grep -q "$ZMS_ENV" && $BIN/sendProcessListToScopeAdmin --commit 141
+        $API config cron queueProcessListToScopeAdmin|grep -q "$ZMS_ENV" && $BIN/queueProcessListToScopeAdmin --commit 141
         $API config cron calculateSlots|grep -q "$ZMS_ENV" && $BIN/calculateSlots --commit --delete
     fi
 fi
