+++ zmsbackend/cron/cronjob.minutly
@@ -13,7 +13,7 @@ if [[ $ZMS_CRONROOT =~ ^1|on|true|yes$ ]]; then
 	      $API config cron deleteUnconfirmedAppointmentData|grep -q "$ZMS_ENV" && $BIN/deleteUnconfirmedAppointmentData --commit
         $API config cron deallocateAppointmentData|grep -q "$ZMS_ENV" && $BIN/deallocateAppointmentData --commit
         $API config cron deleteSessionData|grep -q "$ZMS_ENV" && $BIN/deleteSessionData --commit
-        $API config cron sendMailReminder|grep -q "$ZMS_ENV" && $BIN/sendMailReminder 120 --commit
+        $API config cron queueMailReminder|grep -q "$ZMS_ENV" && $BIN/queueMailReminder 120 --commit
         $API config cron deleteReservedData|grep -q "$ZMS_ENV" && $BIN/deleteReservedData --commit
         $API config cron calculateSlots|grep -q "$ZMS_ENV" && $BIN/calculateSlots --commit
         $API config cron resetApiQuota|grep -q "$ZMS_ENV" && $BIN/resetApiQuota --commit
