+++ zmsclient/tests/mockup/config/config_token.yaml
@@ -8,7 +8,7 @@ response:
     Content-Type: 
     - "application/json"
   body: > 
-   {"$schema":"https://mockup:8083/terminvereinbarung/api/2/","meta":{"$schema":"https://schema.berlin.de/queuemanagement/metaresult.json","error":false,"generated":"2022-04-07T16:47:53+02:00","server":"Zmsbackend-ENV (2.24.05-20-g3df8532)"},"data":{"$schema":"https://schema.berlin.de/queuemanagement/config.json","appointments":{"urlChange":"https://service.berlin.de/terminvereinbarung/termin/manage/","urlAppointments":"https://service.berlin.de/terminvereinbarung/"},"ticketprinter":{"baseUrl":"/terminvereinbarung/ticketprinter/"},"calldisplay":{"baseUrl":"/terminvereinbarung/calldisplay/"},"emergency":{"refreshInterval":"5"},"support":{"eMail":"hotline@itdz-berlin.de","telephone":"(030) 9222-2000"},"sources_dldb_last":"2022-04-07T16:02:08+02:00","cron":{"archiveStatisticData":"stage,dev","calculateDayOffList":"prod,stage,dev","calculateSlots":"www,prod,stage,dev","deallocateAppointmentData":"stage,dev","deleteAppointmentData":"stage,dev","deleteBlacklistedMail":"www,prod,stage,dev","deleteDayoffData":"stage,dev","deleteOldAvailabilityData":"prod,stage,dev","deleteReservedData":"stage,dev","deleteSessionData":"www,prod,stage,dev","migrate":"stage,dev","resetApiQuota":"www,prod,stage,dev","resetGhostWorkstationCount":"www,prod,stage,dev","resetWorkstations":"www,prod,stage,dev","sendMailReminder":"none","sendProcessListToScopeAdmin":"none","updateDldbData":"www,prod,stage,dev"},"setting":{"wsrepsync":"7"},"status":{"calculateSlotsForceVerbose":"0","calculateSlotsLastRun":"2022-04-07 16:45:03","calculateSlotsLastStart":"2022-04-07 16:45:03","calculateSlotsMaintenanceSQL":""}}}
+   {"$schema":"https://mockup:8083/terminvereinbarung/api/2/","meta":{"$schema":"https://schema.berlin.de/queuemanagement/metaresult.json","error":false,"generated":"2022-04-07T16:47:53+02:00","server":"Zmsbackend-ENV (2.24.05-20-g3df8532)"},"data":{"$schema":"https://schema.berlin.de/queuemanagement/config.json","appointments":{"urlChange":"https://service.berlin.de/terminvereinbarung/termin/manage/","urlAppointments":"https://service.berlin.de/terminvereinbarung/"},"ticketprinter":{"baseUrl":"/terminvereinbarung/ticketprinter/"},"calldisplay":{"baseUrl":"/terminvereinbarung/calldisplay/"},"emergency":{"refreshInterval":"5"},"support":{"eMail":"hotline@itdz-berlin.de","telephone":"(030) 9222-2000"},"sources_dldb_last":"2022-04-07T16:02:08+02:00","cron":{"archiveStatisticData":"stage,dev","calculateDayOffList":"prod,stage,dev","calculateSlots":"www,prod,stage,dev","deallocateAppointmentData":"stage,dev","deleteAppointmentData":"stage,dev","deleteBlacklistedMail":"www,prod,stage,dev","deleteDayoffData":"stage,dev","deleteOldAvailabilityData":"prod,stage,dev","deleteReservedData":"stage,dev","deleteSessionData":"www,prod,stage,dev","migrate":"stage,dev","resetApiQuota":"www,prod,stage,dev","resetGhostWorkstationCount":"www,prod,stage,dev","resetWorkstations":"www,prod,stage,dev","queueMailReminder":"none","queueProcessListToScopeAdmin":"none","updateDldbData":"www,prod,stage,dev"},"setting":{"wsrepsync":"7"},"status":{"calculateSlotsForceVerbose":"0","calculateSlotsLastRun":"2022-04-07 16:45:03","calculateSlotsLastStart":"2022-04-07 16:45:03","calculateSlotsMaintenanceSQL":""}}}
 control:
   scenario:
     name: token
